requirements.txt
# Core ML/AI Libraries
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0

# Transformers and NLP
transformers>=4.30.0
datasets>=2.12.0
accelerate>=0.20.0
sentencepiece>=0.1.99
safetensors>=0.3.1

# Computer Vision
opencv-python>=4.7.0
pillow>=9.5.0
ultralytics>=8.0.0  # YOLOv8 for object detection and segmentation

# Detectron2 (install separately - see installation instructions)
# detectron2 @ git+https://github.com/facebookresearch/detectron2.git

# Web Framework
streamlit>=1.25.0

# Utility Libraries
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
requests>=2.31.0
tqdm>=4.65.0

# Development and Testing (optional)
pytest>=7.4.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.4.0

# Additional dependencies for enhanced functionality
scienceplot>=2.1.0  # For scientific plotting
imageio>=2.31.0     # For image I/O operations
scikit-image>=0.21.0  # For advanced image processing